feat: separado camada de logica do primeiro passo
